fortify-tests/pyproject.toml
[tool.poetry]
name = "fortify-tests"
version = "0.1.0"
description = ""
authors = ["Brandon C <brandoncampbell98@hotmail.co.uk>"]
readme = "README.md"
package-mode = false

[tool.pytest.ini_options]
testpaths = [
	"tests/integration",
	"tests/e2e",
	"tests/perf"
]

markers = [
	"healthcheck: mark a test as healthcheck test",
	"e2e: mark a test as end to end test",
	"it: mark a test as integration test"
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
pytest-order = "^1.2.1"
pytest-playwright = "^0.5.1"
pytest-html = "^4.1.1"
locust = "^2.29.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
